requirements.txt
appmap>=1.1.0.dev0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anthropic==0.21.3
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
confection==0.1.4
ConfigArgParse==1.7
cymem==2.0.8
dataclasses-json==0.6.4
deepdiff==6.7.1
deprecation==2.1.0
diff-match-patch==20230430
diskcache==5.6.3
distro==1.9.0
docstring-parser==0.15
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
fastapi==0.110.0
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
gotrue==2.4.2
greenlet==3.0.3
grep-ast==0.2.4
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
hiredis==2.3.2
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
instructor==0.6.7
Jinja2==3.1.4
joblib==1.3.2
jsonpatch==1.33
jsonpath==0.82.2
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kafka-python==2.0.2
langchain-community==0.0.34
langchain-core==0.1.46
langcodes==3.4.0
langsmith==0.1.51
language_data==1.2.0
litellm==1.36.2
marisa-trie==1.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
numpy
openai==1.30.3
ordered-set==4.1.0
orjson==3.10.1
outcome==1.3.0.post0
packaging
pandas==2.2.2
pathspec==0.12.1
phonenumbers==8.13.35
pillow==10.3.0
playwright==1.43.0
pluggy==1.4.0
postgrest==0.16.4
preshed==3.0.9
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
pyee==11.1.0
Pygments==2.18.0
pypandoc==1.13
pyparsing==3.1.2
pyreadline3==3.4.1
PySocks==1.7.1
pytest==8.1.1
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
realtime==1.0.4
redis==5.0.3
redisvl==0.1.2
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
requests-file==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.13.0
selenium==4.20.0
playwright
sentence-transformers
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
srsly==2.4.8
starlette==0.36.3
storage3==0.7.4
streamlit==1.34.0
StrEnum==0.4.15
supabase==2.4.3
supafunc==0.4.5
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
thinc==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tldextract==5.1.2
tokenizers
toml==0.10.2
toolz==0.12.1
torch==2.2.2
tornado==6.4
tqdm==4.66.4
transformers
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
trio==0.25.0
trio-websocket==0.11.1
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
wasabi==1.1.2
watchdog==4.0.0
wcwidth==0.2.13
weasel==0.3.4
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
